#66 move some services to abstractions project
